+++ client/src/index.css
@@ -28,7 +28,7 @@ body {
   line-height: 1.5;
   font-size: 18px;
   letter-spacing: .1em;
-  background: #F4F1DE;
+  background:url("data:image/svg+xml,%3Csvg xmlns='http://www.w3.org/2000/svg' height='100%25' width='100%25'%3E%3Cdefs%3E%3Cpattern id='doodad' width='40' height='40' viewBox='0 0 40 40' patternUnits='userSpaceOnUse' patternTransform='rotate(135)'%3E%3Crect width='100%25' height='100%25' fill='rgba(26, 32, 44,1)'/%3E%3Cpath d='M0 20h15v1h-15zM40 20h15v1h-15zM40 20h15v1h-15z' fill='rgba(10, 147, 150,1)'/%3E%3Cpath d='M20 20h15v1h-15zM60 20h15v1h-15z' fill='rgba(224, 122, 95,1)'/%3E%3C/pattern%3E%3C/defs%3E%3Crect fill='url(%23doodad)' height='200%25' width='200%25'/%3E%3C/svg%3E ");
 }
 
 #adopt-dog {
@@ -42,7 +42,7 @@ body {
     width: fit-content;
     height: fit-content;
     border-radius: 300px;
-    box-shadow: 15px 15px 15px #0A9396;
+    box-shadow: 15px 15px 15px #5B96A9;
     display: flex;
     justify-content: center;
     align-items: center;
@@ -66,7 +66,7 @@ body {
 
   #card-body {
     margin-top: 6%;
-    background-color: #0a9496ad;
+    background-color: #5b96a9f5;
     border-radius: 8px;
     border-color: #E07A5F;
   }
@@ -115,7 +115,11 @@ body {
   }
 
   #start {
-    color: #3D405B;
+    color: #E07A5F;
+  }
+
+  #ready {
+    color:#FFFCF6;
   }
 
   #card-title {
